cleaned up state machine stuff and added a gitignore
